+++ contexts/LocaleContext.ts
@@ -1,7 +1,7 @@
 import { createContext } from "react";
 import { OpenModal, SharedPageData, FolderPageData } from "../types/common";
 
-export type LocaleContextType = SharedPageData[] | FolderPageData[];
+export type LocaleContextType = SharedPageData | FolderPageData[] | SharedPageData[];
 
 export const LocaleContext = createContext<LocaleContextType>([]);
 export const ModalContext = createContext<OpenModal>();
