+++ pages/api/GetData.tsx
@@ -1,52 +1,7 @@
-import axios from "axios";
 import API_BASE_URL from "../../constants";
+import axiosInstance from "./axios";
 
-interface SharedFolderLink {
-  id: number;
-  createdAt: string;
-  url: string;
-  title: string;
-  description: string;
-  imageSource: string;
-}
-
-interface SharedFolder {
-  folder: {
-    id: number;
-    name: string;
-    owner: {
-      id: number;
-      name: string;
-      profileImageSource: string;
-    };
-    links: SharedFolderLink[];
-  };
-}
-
-const convertToSnakeCase = (data: SharedFolder) => {
-  const { links } = data.folder;
-  const convertedLinks = links.map((link) => {
-    const convertedLink = Object.fromEntries(
-      Object.entries(link).map(([key, value]) => [
-        key.replace(/[A-Z]/g, (match) => `_${match.toLowerCase()}`),
-        value,
-      ])
-    );
-    return convertedLink;
-  });
-  return convertedLinks;
-};
-
-const axiosInstance = axios.create();
-
-axiosInstance.interceptors.response.use((response) => {
-  if (response.data && response.data.folder) {
-    response.data.folder.links = convertToSnakeCase(response.data);
-  }
-  return response;
-});
-
-const getData = async (url: string) => {
+const requestApi = async (url: string) => {
   try {
     const response = await axiosInstance.get(`${API_BASE_URL}${url}`);
 
@@ -61,4 +16,4 @@ const getData = async (url: string) => {
   }
 };
 
-export default getData;
+export default requestApi;
